:bug: 방안에 유저가 있는경우 방삭제불가 서버에러처리용

+++ src/room/room.service.ts
@@ -158,9 +158,8 @@ export class RoomService {
       where: { uuid: uuid },
     });
     if (!findRoom) throw RoomException.roomNotFound();
-
     //방안에 유저가 있는 경우 에러띄움
-    //if (findRoom.count) throw RoomException.roomUserexists();
+    if (findRoom.count) throw RoomException.roomUserexists();
     if (userId != findRoom.ownerId) throw UserException.userUnauthorized();
     const deleteResult = await this.prisma.room.delete({
       where: { uuid: uuid },
